+++ components/event/event-listener/src/test/java/uk/gov/justice/services/event/buffer/it/EventBufferAndFilterChainIT.java
@@ -27,7 +27,6 @@ import uk.gov.justice.services.core.cdi.LoggerProducer;
 import uk.gov.justice.services.core.dispatcher.DispatcherCache;
 import uk.gov.justice.services.core.dispatcher.DispatcherFactory;
 import uk.gov.justice.services.core.dispatcher.EmptySystemUserProvider;
-import uk.gov.justice.services.core.dispatcher.RequesterProducer;
 import uk.gov.justice.services.core.dispatcher.ServiceComponentObserver;
 import uk.gov.justice.services.core.dispatcher.SystemUserUtil;
 import uk.gov.justice.services.core.envelope.EnvelopeValidationExceptionHandlerProducer;
@@ -42,11 +41,9 @@ import uk.gov.justice.services.core.interceptor.InterceptorChainObserver;
 import uk.gov.justice.services.core.interceptor.InterceptorChainProcessor;
 import uk.gov.justice.services.core.interceptor.InterceptorChainProcessorProducer;
 import uk.gov.justice.services.core.interceptor.InterceptorChainProvider;
-import uk.gov.justice.services.core.jms.DefaultJmsDestinations;
-import uk.gov.justice.services.core.jms.JmsSenderFactory;
 import uk.gov.justice.services.core.json.DefaultJsonSchemaValidator;
 import uk.gov.justice.services.core.json.JsonSchemaLoader;
-import uk.gov.justice.services.core.sender.ComponentDestination;
+import uk.gov.justice.services.core.requester.RequesterProducer;
 import uk.gov.justice.services.core.sender.SenderProducer;
 import uk.gov.justice.services.event.buffer.EventBufferInterceptor;
 import uk.gov.justice.services.event.buffer.core.service.ConsecutiveEventBufferService;
@@ -116,10 +113,7 @@ public class EventBufferAndFilterChainIT {
             ServiceComponentObserver.class,
 
             SenderProducer.class,
-            JmsSenderFactory.class,
-            ComponentDestination.class,
             DefaultJmsEnvelopeSender.class,
-            DefaultJmsDestinations.class,
             DefaultEnvelopeConverter.class,
 
             StringToJsonObjectConverter.class,