+++ core/src/main/java/uk/gov/justice/services/core/requester/RequesterProducer.java
@@ -1,6 +1,9 @@
-package uk.gov.justice.services.core.dispatcher;
+package uk.gov.justice.services.core.requester;
 
 import uk.gov.justice.services.core.annotation.ServiceComponent;
+import uk.gov.justice.services.core.dispatcher.DispatcherCache;
+import uk.gov.justice.services.core.dispatcher.DispatcherDelegate;
+import uk.gov.justice.services.core.dispatcher.SystemUserUtil;
 import uk.gov.justice.services.core.envelope.EnvelopeValidationExceptionHandler;
 import uk.gov.justice.services.core.envelope.EnvelopeValidator;
 import uk.gov.justice.services.core.json.JsonSchemaValidator;