+++ example-context/example-service/example-query/example-query-api/src/main/java/uk/gov/justice/services/example/cakeshop/query/api/CakeOrdersQueryApi.java
@@ -5,7 +5,7 @@ import static uk.gov.justice.services.core.annotation.Component.QUERY_API;
 
 import uk.gov.justice.services.core.annotation.Handles;
 import uk.gov.justice.services.core.annotation.ServiceComponent;
-import uk.gov.justice.services.core.dispatcher.Requester;
+import uk.gov.justice.services.core.requester.Requester;
 import uk.gov.justice.services.messaging.JsonEnvelope;
 
 import javax.inject.Inject;