+++ example-context/example-service/example-query/example-query-controller/src/main/java/uk/gov/justice/services/example/cakeshop/query/controller/RecipesQueryController.java
@@ -4,7 +4,7 @@ import static uk.gov.justice.services.core.annotation.Component.QUERY_CONTROLLER
 
 import uk.gov.justice.services.core.annotation.Handles;
 import uk.gov.justice.services.core.annotation.ServiceComponent;
-import uk.gov.justice.services.core.dispatcher.Requester;
+import uk.gov.justice.services.core.requester.Requester;
 import uk.gov.justice.services.messaging.JsonEnvelope;
 
 import javax.inject.Inject;