+++ framework-api/framework-api-core/src/main/java/uk/gov/justice/services/core/sender/SenderFactory.java
@@ -1,6 +1,0 @@
-package uk.gov.justice.services.core.sender;
-
-public interface SenderFactory {
-
-    Sender createSender(final String componentDestination);
-}