+++ test-utils/test-utils-core/src/main/java/uk/gov/justice/services/test/utils/core/helper/ServiceComponents.java
@@ -11,7 +11,7 @@ import static uk.gov.justice.services.messaging.JsonObjectMetadata.metadataWithD
 
 import uk.gov.justice.services.core.annotation.Handles;
 import uk.gov.justice.services.core.annotation.ServiceComponent;
-import uk.gov.justice.services.core.dispatcher.Requester;
+import uk.gov.justice.services.core.requester.Requester;
 import uk.gov.justice.services.core.sender.Sender;
 import uk.gov.justice.services.messaging.JsonEnvelope;
 import uk.gov.justice.services.test.utils.core.matchers.HandlerClassMatcher;