+++ test-utils/test-utils-core/src/main/java/uk/gov/justice/services/test/utils/core/mock/JsonSchemaValidatingMockMaker.java
@@ -3,7 +3,7 @@ package uk.gov.justice.services.test.utils.core.mock;
 import static java.util.Optional.empty;
 
 import uk.gov.justice.services.common.converter.jackson.ObjectMapperProducer;
-import uk.gov.justice.services.core.dispatcher.Requester;
+import uk.gov.justice.services.core.requester.Requester;
 import uk.gov.justice.services.core.envelope.EnvelopeValidator;
 import uk.gov.justice.services.core.envelope.RethrowingValidationExceptionHandler;
 import uk.gov.justice.services.core.json.DefaultJsonSchemaValidator;