+++ test-utils/test-utils-core/src/test/java/uk/gov/justice/services/test/utils/core/matchers/HandlerMethodMatcherTest.java
@@ -7,7 +7,7 @@ import static uk.gov.justice.services.test.utils.core.matchers.HandlerMethodMatc
 
 import uk.gov.justice.services.core.annotation.Handles;
 import uk.gov.justice.services.core.annotation.ServiceComponent;
-import uk.gov.justice.services.core.dispatcher.Requester;
+import uk.gov.justice.services.core.requester.Requester;
 import uk.gov.justice.services.core.sender.Sender;
 import uk.gov.justice.services.messaging.JsonEnvelope;